--- src/app/rankings/page.tsx
@@ -1,9 +1,0 @@
-import type { Metadata } from "next";
-
-export const metadata: Metadata = {
-  title: "Rankings",
-};
-
-const Page = () => <div>Rankings page</div>;
-
-export default Page;
