+++ src/components/NavigationBar/OPTION_LIST.ts
@@ -1,6 +1,6 @@
 export const OPTION_LIST = [
   ["Problems", "/problems", "Probs"],
   ["Submissions", "/submissions", "Subs"],
-  ["Rankings", "/rankings", "Ranks"],
+  ["Rankings", "/rankings/0/1", "Ranks"],
   ["About", "/about", "About"],
 ] as const;
