+++ src/shared/schemas/loginSchema.ts
@@ -7,18 +7,4 @@ export const loginSchema = z.object({
 
 export type LoginSchemaType = z.infer<typeof loginSchema>;
 
-export interface UserData {
-  id: number;
-  username: string;
-  name: string;
-  group: {
-    id: number;
-    name: string;
-  };
-  status: string;
-  avatar: string;
-  score: number;
-  solved: number;
-  ratio: number;
-  email: string;
-}
+export type { UserData } from "@/shared/types";
